+++ oscar-cbls/src/main/scala/oscar/cbls/business/routing/package.scala
@@ -42,23 +42,23 @@ package object routing
   with TwoOptAPI {
 
   type VRP = oscar.cbls.business.routing.model.VRP
-  def vrp(m: Store,n:Int, v: Int, maxPivotPerValuePercent: Long = 4L): VRP =
+  def vrp(m: Store,n:Int, v: Int, maxPivotPerValuePercent: Int = 4): VRP =
     new VRP(m, n, v, maxPivotPerValuePercent)
 
   type TTFMatrix = oscar.cbls.business.routing.model.TTFMatrix
-  def ttfMatrix(nodeCount: Long, defaultTTF: PrimitiveTravelTimeFunction): TTFMatrix =
+  def ttfMatrix(nodeCount: Int, defaultTTF: PrimitiveTravelTimeFunction): TTFMatrix =
     new TTFMatrix(nodeCount, defaultTTF)
 
   type TTFConst = oscar.cbls.business.routing.model.TTFConst
-  def ttfConst(travelDuration: Long): TTFConst =
+  def ttfConst(travelDuration: Int): TTFConst =
     new TTFConst(travelDuration)
 
   type TTFHistogram = oscar.cbls.business.routing.model.TTFHistogram
-  def ttfHistogram(nbSlots: Long, overallDuration: Long): TTFHistogram =
+  def ttfHistogram(nbSlots: Int, overallDuration: Int): TTFHistogram =
     new TTFHistogram(nbSlots, overallDuration)
 
   type TTFSegments = oscar.cbls.business.routing.model.TTFSegments
-  def ttfSegments(nbPoints: Long, overallDuration: Long): TTFSegments =
+  def ttfSegments(nbPoints: Int, overallDuration: Int): TTFSegments =
     new TTFSegments(nbPoints, overallDuration)
 
   type PrimitiveTravelTimeFunction = oscar.cbls.business.routing.model.PrimitiveTravelTimeFunction
